requirements.txt
requests==2.31.0
python-dotenv==1.0.0
PyQt6==6.9.0
lxml==4.9.1
tqdm==4.66.1
pydub==0.25.1
numpy==1.24.1
librosa==0.11.0
soundfile==0.13.1
ffmpeg-python==0.2.0
future==1.0.0
